Apply same changes to ServiceUiApp
