+++ packages/frinx-workflow-ui/src/pages/workflowList/WorkflowListHeader.js
@@ -1,41 +1,10 @@
 // @flow
-import React, { useContext, useEffect } from 'react';
+import React, { useContext } from 'react';
 import { Button } from 'react-bootstrap';
 import { HttpClient as http } from '../../common/HttpClient';
-import { exportButton } from './workflowUtils';
 import { GlobalContext } from '../../common/GlobalContext';
 import { useRouteMatch, useHistory } from 'react-router-dom';
 import PageContainer from '../../common/PageContainer';
-
-const workflowModifyButtons = (openFileUpload, history, frontendUrlPrefix) => {
-  console.log(frontendUrlPrefix)
-  return [
-    <Button
-      key="builder-btn"
-      variant="outline-primary"
-      style={{ marginLeft: '30px' }}
-      onClick={() => history.push(frontendUrlPrefix + '/builder')}
-    >
-      <i className="fas fa-plus" />
-      &nbsp;&nbsp;New
-    </Button>,
-    <Button key="import-btn" variant="outline-primary" style={{ marginLeft: '5px' }} onClick={openFileUpload}>
-      <i className="fas fa-file-import" />
-      &nbsp;&nbsp;Import
-    </Button>,
-  ];
-};
-
-const upperMenu = (history, openFileUpload, backendApiUrlPrefix, frontendUrlPrefix) => {
-  return (
-    <h1 style={{ marginBottom: '20px' }}>
-      <i style={{ color: 'grey' }} className="fas fa-cogs" />
-      &nbsp;&nbsp;Workflows
-      {workflowModifyButtons(openFileUpload, history, frontendUrlPrefix)}
-      {exportButton({ backendApiUrlPrefix: backendApiUrlPrefix })}
-    </h1>
-  );
-};
 
 const WorkflowListHeader = () => {
   const global = useContext(GlobalContext);
@@ -71,11 +40,49 @@ const WorkflowListHeader = () => {
     document.getElementById('upload-files').addEventListener('change', importFiles);
   };
 
-  let menu = upperMenu(history, openFileUpload, global.backendApiUrlPrefix, path);
+  const exportFile = backendApiUrlPrefix => {
+    http.get(backendApiUrlPrefix + '/metadata/workflow').then(res => {
+      const zip = new JSZip();
+      let workflows = res.result || [];
+
+      workflows.forEach(wf => {
+        zip.file(wf.name + '.json', JSON.stringify(wf, null, 2));
+      });
+
+      zip.generateAsync({ type: 'blob' }).then(function(content) {
+        saveAs(content, 'workflows.zip');
+      });
+    });
+  };
 
   return (
     <PageContainer>
-      {menu}
+      <h1 style={{ marginBottom: '20px' }}>
+        <i style={{ color: 'grey' }} className="fas fa-cogs" />
+        &nbsp;&nbsp;Workflows
+        <Button
+          key="builder-btn"
+          variant="outline-primary"
+          style={{ marginLeft: '30px' }}
+          onClick={() => history.push(path + '/builder')}
+        >
+          <i className="fas fa-plus" />
+          &nbsp;&nbsp;New
+        </Button>
+        <Button key="import-btn" variant="outline-primary" style={{ marginLeft: '5px' }} onClick={openFileUpload}>
+          <i className="fas fa-file-import" />
+          &nbsp;&nbsp;Import
+        </Button>
+        <Button
+          key="export-btn"
+          variant="outline-primary"
+          style={{ marginLeft: '5px' }}
+          onClick={() => exportFile(global.backendApiUrlPrefix)}
+        >
+          <i className="fas fa-file-export" />
+          &nbsp;&nbsp;Export
+        </Button>
+      </h1>
       <input id="upload-files" multiple type="file" hidden />
     </PageContainer>
   );
